Added features to compileJSP server. Working on tests
